auth_service/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
django-cors-headers==3.2.0
